refactor database entities

+++ app/src/main/kotlin/ru/cookedapp/cooked/data/db/entity/Ingredient.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Ingredient(
-    var productId: Int,
+    var productId: Long,
     var amount: Float,
-    var amountUnitId: Int
+    var amountUnitId: Long,
 )
