+++ app/src/main/kotlin/ru/cookedapp/cooked/data/repository/ProductRepository.kt
@@ -22,19 +22,19 @@ class ProductRepository(private val productDao: ProductDao) {
         productDao.delete(product)
     }
 
-    suspend fun removeById(id: Int) = withContext(Dispatchers.IO) {
+    suspend fun removeById(id: Long) = withContext(Dispatchers.IO) {
         productDao.deleteById(id)
     }
 
-    suspend fun increaseUsage(id: Int) = withContext(Dispatchers.IO) {
+    suspend fun increaseUsage(id: Long) = withContext(Dispatchers.IO) {
         productDao.increaseUsages(id)
     }
 
-    suspend fun decreaseUsage(id: Int) = withContext(Dispatchers.IO) {
+    suspend fun decreaseUsage(id: Long) = withContext(Dispatchers.IO) {
         productDao.decreaseUsages(id)
     }
 
-    suspend fun renameProduct(id: Int, name: String) = withContext(Dispatchers.IO) {
+    suspend fun renameProduct(id: Long, name: String) = withContext(Dispatchers.IO) {
         productDao.renameProduct(id, name)
     }
 }
